split serialization header
